+++ include/sensors/firmware/pressure_sensor_callbacks.h
@@ -1,6 +1,11 @@
 #include "platform_specific_hal_conf.h"
+#include "sensors/core/tasks/pressure_driver.hpp"
 
-typedef void(*sensor_callback) ();
-sensor_callback data_ready_callback;
-//static void * data_ready_callback;
-void drdy_callback();
+void HAL_GPIO_EXTI_Falling_Callback(uint16_t GPIO_Pin) {
+    if (GPIO_Pin == GPIO_PIN_8 ||
+        GPIO_Pin == GPIO_PIN_9 ||
+        GPIO_Pin == GPIO_PIN_15)
+    {
+        MMR92C04::drdy_interrupt_response();
+    }
+}
